+++ src/events/client/interactionCreate.js
@@ -2,76 +2,29 @@ const Discord = require('discord.js');
 const Captcha = require("@haileybot/captcha-generator");
 
 const reactionSchema = require("../../database/models/reactionRoles");
-const banSchema = require("../../database/models/userBans");
 const verify = require("../../database/models/verify");
 const Commands = require("../../database/models/customCommand");
 const CommandsSchema = require("../../database/models/customCommandAdvanced");
+
 module.exports = async (client, interaction) => {
     // Commands
     if (interaction.isCommand() || interaction.isUserContextMenuCommand()) {
-        banSchema.findOne({ User: interaction.user.id }, async (err, data) => {
-            if (data) {
-                return client.errNormal({
-                    error: "You have been banned by the developers of this bot",
-                    type: 'ephemeral'
-                }, interaction);
-            }
-            else {
-                const cmd = client.commands.get(interaction.commandName);
-                if (!cmd){
-                    const cmdd = await Commands.findOne({
-                        Guild: interaction.guild.id,
-                        Name: interaction.commandName,
-                    });
-                    if (cmdd) {
-                        return interaction.channel.send({ content: cmdd.Responce });
-                    }
+        // Declare the variable `cmd` before using it
+        const cmd = client.commands.get(interaction.commandName);
 
-                    const cmdx = await CommandsSchema.findOne({
-                        Guild: interaction.guild.id,
-                        Name: interaction.commandName,
-                    });
-                    if (cmdx) {
-                        // Remove interaction
-                        if (cmdx.Action == "Normal") {
-                            return interaction.reply({ content: cmdx.Responce });
-                        } else if (cmdx.Action == "Embed") {
-                            return client.simpleEmbed(
-                                {
-                                    desc: `${cmdx.Responce}`,
-                                    type: 'reply'
-                                },
-                                interaction,
-                            );
-                        } else if (cmdx.Action == "DM") {
-                            await interaction.deferReply({ ephemeral: true });
-                            interaction.editReply({ content: "I have sent you something in your DMs" });
-                            return interaction.user.send({ content: cmdx.Responce }).catch((e) => {
-                                client.errNormal(
-                                    {
-                                        error: "I can't DM you, maybe you have DM turned off!",
-                                        type: 'ephemeral'
-                                    },
-                                    interaction,
-                                );
-                            });
-                        }
-                    }
-                }
-                if (interaction.options._subcommand !== null && interaction.options.getSubcommand() == "help") {
-                    const commands = interaction.client.commands.filter(x => x.data.name == interaction.commandName).map((x) => x.data.options.map((c) => '`' + c.name + '` - ' + c.description).join("\n"));
+        // Remove interaction
+        if (interaction.options._subcommand !== null && interaction.options.getSubcommand() == "help") {
+            const commands = interaction.client.commands.filter(x => x.data.name == interaction.commandName).map((x) => x.data.options.map((c) => '`' + c.name + '` - ' + c.description).join("\n"));
 
-                    return client.embed({
-                        title: `❓・Help panel`,
-                        desc: `Get help with the commands in \`${interaction.commandName}\` \n\n${commands}`,
-                        type: 'reply'
-                    }, interaction)
-                }
+            return client.embed({
+                title: `❓・Help panel`,
+                desc: `Get help with the commands in \`<span class="math-inline">\{interaction\.commandName\}\\\` \\n\\n</span>{commands}`,
+                type: 'reply'
+            }, interaction)
+        }
 
-                if(cmd) cmd.run(client, interaction, interaction.options._hoistedOptions).catch(err => {
-                    client.emit("errorCreate", err, interaction.commandName, interaction)
-                })
-            }
+        if(cmd) cmd.run(client, interaction, interaction.options._hoistedOptions).catch(err => {
+            client.emit("errorCreate", err, interaction.commandName, interaction)
         })
     }